カードコンポーネント細分化
