+++ pages/show/index.tsx
@@ -2,9 +2,9 @@ import { NextPage } from "next";
 import { useContext } from "react";
 import { GeolocationContext, ViewIdContext } from "../_app";
 import Modal from "../../src/components/Modal";
-import IntroductionCard from "../../src/components/Card/IntroductionCard";
 import Gallery from "../../src/components/Show/Gallery";
 import Header from "../../src/components/Header";
+import ShowIntroduction from "../../src/components/Show/ShowIntroduction";
 
 const Show: NextPage = () => {
   const { position } = useContext(GeolocationContext);
@@ -17,7 +17,7 @@ const Show: NextPage = () => {
       <div className="bg-white py-6 sm:py-8 lg:py-12">
         {viewId ? (
           <Modal closeModal={() => setViewId(null)}>
-            <IntroductionCard />
+            <ShowIntroduction />
           </Modal>
         ) : null}
